blog-ai/soundtrack_rating/requirements.txt
kafka-python
joblib
numpy
pandas
scikit-learn
elasticsearch
spotipy
google-api-python-client